backpy/requirements.txt
Flask==2.0.2
flask-cors==3.0.10
Pillow==8.1.0  # For image processing